+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.0 (June 11, 2018)
+ * saved application url in window location for easy linking
+ * auto loaded when using iframe mode
+ * removed refresh/auto and sync state automatically
+ * added ability for devtool to move history backwards and forwards
+
 ## 2.0.0 (June 11, 2018)
  * added ability to open in window, since iframes cannot render http on https page, which is needed for github pages